DNAbin2indel {ape}	R Documentation

Recode Blocks of Indels

Description

This function scans a set of aligned DNA sequences and returns a matrix with information of the localisations and lengths on alignment gaps.

Usage

DNAbin2indel(x)

Arguments

x
an object of class "DNAbin".

Details

The output matrix has the same dimensions than the input one with, either a numeric value where an alignment gap starts giving the length of the gap, or zero. The rownames are kept.

Value

a numeric matrix.

Author(s)

Emmanuel Paradis

See Also

DNAbin, as.DNAbin, del.gaps, seg.sites, image.DNAbin, checkAlignment

[Package ape version 5.2 Index]