Initialize.corPhyl {ape}	R Documentation

Initialize a ‘corPhyl’ Structure Object

Description

Initialize a corPhyl correlation structure object. Does the same as Initialize.corStruct, but also checks the row names of data and builds an index.

Usage

	## S3 method for class 'corPhyl'
Initialize(object, data, ...)

Arguments

object
An object inheriting from class corPhyl.

data
The data to use. If it contains rownames, they are matched with the tree tip labels, otherwise data are supposed to be in the same order than tip labels and a warning is sent.

...
some methods for this generic require additional arguments. None are used in this method.

Value

An initialized object of same class as object.

Author(s)

Julien Dutheil julien.dutheil@univ-montp2.fr

See Also

corClasses, Initialize.corStruct.

[Package ape version 5.2 Index]