additive {ape}	R Documentation

Incomplete Distance Matrix Filling

Description

Fills missing entries from incomplete distance matrix using the additive or the ultrametric procedure (see reference for details).

Usage

additive(X)
ultrametric(X)

Arguments

X
a distance matrix or an object of class "dist".

Value

a distance matrix.

Author(s)

Andrei Popescu niteloserpopescu@gmail.com

References

Makarenkov, V. and Lapointe, F.-J. (2004) A weighted least-squares approach for inferring phylogenies from incomplete distance matrices. Bioinformatics, 20, 2113–2121.

[Package ape version 5.2 Index]