alex {ape}	R Documentation

Alignment Explorer With Multiple Devices

Description

This function helps to explore DNA alignments by zooming in. The user clicks twice defining the opposite corners of the portion which is extracted and drawned on a new window.

Usage

alex(x, ...)

Arguments

x
an object of class "DNAbin".

...
further arguments to pass to image.DNAbin.

Details

This function works with a DNA alignment (freshly) plotted on an interactive graphical device (i.e., not a file) with image. After calling alex, the user clicks twice defining a rectangle in the alignment, then this portion of the alignment is extacted and plotted on a new window. The user can click as many times on the alignment. The process is stopped by a right-click. If the user clicks twice outside the alignment, a message “Try again!” is printed.

Each time alex is called, the alignment is plotted on a new window without closing or deleting those possibly already plotted.

In all cases, the device where x is plotted is the active window after the operation. It should not be closed during the whole process.

Value

NULL

Author(s)

Emmanuel Paradis

See Also

image.DNAbin, trex, alview

Examples

## Not run: 
data(woodmouse)
image(woodmouse)
alex(woodmouse)

## End(Not run)

[Package ape version 5.2 Index]