as.alignment {ape}	R Documentation

Conversion Among DNA Sequence Internal Formats

Description

These functions transform a set of DNA sequences among various internal formats.

Usage

as.alignment(x)
as.DNAbin(x, ...)

## S3 method for class 'character'
as.DNAbin(x, ...)

## S3 method for class 'list'
as.DNAbin(x, ...)

## S3 method for class 'alignment'
as.DNAbin(x, ...)

## S3 method for class 'DNAString'
as.DNAbin(x, ...)

## S3 method for class 'DNAStringSet'
as.DNAbin(x, ...)

## S3 method for class 'PairwiseAlignmentsSingleSubject'
as.DNAbin(x, ...)

## S3 method for class 'DNAMultipleAlignment'
as.DNAbin(x, ...)

## S3 method for class 'DNAbin'
as.character(x, ...)

Arguments

x
a matrix or a list containing the DNA sequences, or an object of class "alignment".

...
further arguments to be passed to or from other methods.

Details

For as.alignment, the sequences given as argument should be stored as matrices or lists of single-character strings (the format used in ape before version 1.10). The returned object is in the format used in the package seqinr to store aligned sequences.

as.DNAbin is a generic function with methods so that it works with sequences stored into vectors, matrices, or lists. It can convert some S4 classes from the package Biostrings in BioConductor. For consistency within ape, this uses an S3-style syntax. To convert objects of class "DNAStringSetList", see the examples.

as.character is a generic function: the present method converts objects of class "DNAbin" into the format used before ape 1.10 (matrix of single characters, or list of vectors of single characters). This function must be used first to convert objects of class "DNAbin" into the class "alignment".

Value

an object of class "alignment" in the case of "as.alignment"; an object of class "DNAbin" in the case of "as.DNAbin"; a matrix of mode character or a list containing vectors of mode character in the case of "as.character".

Author(s)

Emmanuel Paradis

See Also

DNAbin, read.dna, read.GenBank, write.dna

Examples

data(woodmouse)
x <- as.character(woodmouse)
x[, 1:20]
str(as.alignment(x))
identical(as.DNAbin(x), woodmouse)
### conversion from BioConductor:
## Not run: 
if (require(Biostrings)) {
data(phiX174Phage)
X <- as.DNAbin(phiX174Phage)
## base frequencies:
base.freq(X) # from ape
alphabetFrequency(phiX174Phage) # from Biostrings
### for objects of class "DNAStringSetList"
X <- lapply(x, as.DNAbin) # a list of lists
### to put all sequences in a single list:
X <- unlist(X, recursive = FALSE)
class(X) <- "DNAbin"
}

## End(Not run)

[Package ape version 5.2 Index]