axisPhylo {ape}	R Documentation

Axis on Side of Phylogeny

Description

This function adds a scaled axis on the side of a phylogeny plot.

Usage

axisPhylo(side = 1, root.time = NULL, backward = TRUE, ...)

Arguments

side
a numeric value specifying the side where the axis is plotted: 1: below, 2: left, 3: above, 4: right.

root.time
the time assigned to the root node of the tree. By default, this is taken from the root.time element of the tree. If it is absent, this is determined from the next option.

backward
a logical value; if TRUE, the most distant tip from the root is considered as the origin of the time scale; if FALSE, this is the root node.

...
further arguments to be passed to axis.

Details

The further arguments (...) are used to format the axis. They may be font, cex, col, las, and so on (see the help pages on axis and par).

Author(s)

Emmanuel Paradis

See Also

plot.phylo, add.scale.bar, axis, par

Examples

tr <- rtree(30)
ch <- rcoal(30)
plot(ch)
axisPhylo()
plot(tr, "c", FALSE, direction = "u")
axisPhylo(2, las = 1)

[Package ape version 5.2 Index]