bd.ext {ape}	R Documentation

Extended Version of the Birth-Death Models to Estimate Speciation and Extinction Rates

Description

This function fits by maximum likelihood a birth-death model to the combined phylogenetic and taxonomic data of a given clade. The phylogenetic data are given by a tree, and the taxonomic data by the number of species for the its tips.

Usage

bd.ext(phy, S, conditional = TRUE)

Arguments

phy
an object of class "phylo".

S
a numeric vector giving the number of species for each tip.

conditional
whether probabilities should be conditioned on no extinction (mainly to compare results with previous analyses; see details).

Details

A re-parametrization of the birth-death model studied by Kendall (1948) so that the likelihood has to be maximized over d/b and b - d, where b is the birth rate, and d the death rate.

The standard-errors of the estimated parameters are computed using a normal approximation of the maximum likelihood estimates.

If the argument S has names, then they are matched to the tip labels of phy. The user must be careful here since the function requires that both series of names perfectly match, so this operation may fail if there is a typing or syntax error. If both series of names do not match, the values S are taken to be in the same order than the tip labels of phy, and a warning message is issued.

Note that the function does not check that the tree is effectively ultrametric, so if it is not, the returned result may not be meaningful.

If conditional = TRUE, the probabilities of the taxonomic data are calculated conditioned on no extinction (Rabosky et al. 2007). In previous versions of the present function (until ape 2.6-1), unconditional probabilities were used resulting in underestimated extinction rate. Though it does not make much sense to use conditional = FALSE, this option is provided to compare results from previous analyses: if the species richnesses are relatively low, both versions will give similar results (see examples).

Author(s)

Emmanuel Paradis

References

Paradis, E. (2003) Analysis of diversification: combining phylogenetic and taxonomic data. Proceedings of the Royal Society of London. Series B. Biological Sciences, 270, 2499–2505.

Rabosky, D. L., Donnellan, S. C., Talaba, A. L. and Lovette, I. J. (2007) Exceptional among-lineage variation in diversification rates during the radiation of Australia's most diverse vertebrate clade. Proceedings of the Royal Society of London. Series B. Biological Sciences, 274, 2915–2923.

See Also

birthdeath, branching.times, diversi.gof, diversi.time, ltt.plot, yule, yule.cov, bd.time

Examples

### An example from Paradis (2003) using the avian orders:
data(bird.orders)
### Number of species in each order from Sibley and Monroe (1990):
S <- c(10, 47, 69, 214, 161, 17, 355, 51, 56, 10, 39, 152,
       6, 143, 358, 103, 319, 23, 291, 313, 196, 1027, 5712)
bd.ext(bird.orders, S)
bd.ext(bird.orders, S, FALSE) # same than older versions

[Package ape version 5.2 Index]