bird.orders {ape}	R Documentation

Phylogeny of the Orders of Birds From Sibley and Ahlquist

Description

This data set describes the phylogenetic relationships of the orders of birds as reported by Sibley and Ahlquist (1990). Sibley and Ahlquist inferred this phylogeny from an extensive number of DNA/DNA hybridization experiments. The “tapestry” reported by these two authors (more than 1000 species out of the ca. 9000 extant bird species) generated a lot of debates.

The present tree is based on the relationships among orders. The branch lengths were calculated from the values of Delta T50H as found in Sibley and Ahlquist (1990, fig. 353).

Usage

data(bird.orders)

Format

The data are stored as an object of class "phylo" which structure is described in the help page of the function read.tree.

Source

Sibley, C. G. and Ahlquist, J. E. (1990) Phylogeny and classification of birds: a study in molecular evolution. New Haven: Yale University Press.

See Also

read.tree, bird.families

Examples

data(bird.orders)
plot(bird.orders)

[Package ape version 5.2 Index]