birthdeath {ape}	R Documentation

Estimation of Speciation and Extinction Rates With Birth-Death Models

Description

This function fits by maximum likelihood a birth-death model to the branching times computed from a phylogenetic tree using the method of Nee et al. (1994).

Usage

birthdeath(phy)
## S3 method for class 'birthdeath'
print(x, ...)

Arguments

phy
an object of class "phylo".

x
an object of class "birthdeath".

...
further arguments passed to the print function.

Details

Nee et al. (1994) used a re-parametrization of the birth-death model studied by Kendall (1948) so that the likelihood has to be maximized over d/b and b - d, where b is the birth rate, and d the death rate. This is the approach used by the present function.

This function computes the standard-errors of the estimated parameters using a normal approximations of the maximum likelihood estimates: this is likely to be inaccurate because of asymmetries of the likelihood function (Nee et al. 1995). In addition, 95 intervals of both parameters are computed using profile likelihood: they are particularly useful if the estimate of d/b is at the boundary of the parameter space (i.e. 0, which is often the case).

Note that the function does not check that the tree is effectively ultrametric, so if it is not, the returned result may not be meaningful.

Value

An object of class "birthdeath" which is a list with the following components:

tree
the name of the tree analysed.

N
the number of species.

dev
the deviance (= -2 log lik) at its minimum.

para
the estimated parameters.

se
the corresponding standard-errors.

CI
the 95% profile-likelihood confidence intervals.

Author(s)

Emmanuel Paradis

References

Kendall, D. G. (1948) On the generalized “birth-and-death” process. Annals of Mathematical Statistics, 19, 1–15.

Nee, S., May, R. M. and Harvey, P. H. (1994) The reconstructed evolutionary process. Philosophical Transactions of the Royal Society of London. Series B. Biological Sciences, 344, 305–311.

Nee, S., Holmes, E. C., May, R. M. and Harvey, P. H. (1995) Estimating extinctions from molecular phylogenies. in Extinction Rates, eds. Lawton, J. H. and May, R. M., pp. 164–182, Oxford University Press.

See Also

branching.times, diversi.gof, diversi.time, ltt.plot, yule, bd.ext, yule.cov, bd.time

[Package ape version 5.2 Index]