branching.times {ape}	R Documentation

Branching Times of a Phylogenetic Tree

Description

This function computes the branching times of a phylogenetic tree, that is the distance from each node to the tips, under the assumption that the tree is ultrametric. Note that the function does not check that the tree is effectively ultrametric, so if it is not, the returned result may not be meaningful.

Usage

branching.times(phy)

Arguments

phy
an object of class "phylo".

Value

a numeric vector with the branching times. If the phylogeny phy has an element node.label, this is used as names for the returned vector; otherwise the numbers (of mode character) of the matrix edge of phy are used as names.

Author(s)

Emmanuel Paradis

See Also

is.ultrametric

[Package ape version 5.2 Index]