carnivora {ape}	R Documentation

Carnivora body sizes and life history traits

Description

Dataset adapted from Gittleman (1986), including 2 morphological variables (body and brain sizes), 8 life history traits variables and 4 taxonomic variables.

Usage

data(carnivora)

Format

A data frame with 112 observations on 17 variables.

[,1] 	 Order 	 factor 	 Carnivora order

[,2] 	 SuperFamily 	 factor 	 Super family (Caniformia or Feliformia)

[,3] 	 Family 	 factor 	 Carnivora family

[,4] 	 Genus 	 factor 	 Carnivora genus

[,5] 	 Species 	 factor 	 Carnivora species

[,6] 	 FW 	 numeric 	 Female body weight (kg)

[,7] 	 SW 	 numeric 	 Average body weight of adult male and adult female (kg)

[,8] 	 FB 	 numeric 	 Female brain weight (g)

[,9] 	 SB 	 numeric 	 Average brain weight of adult male and adult female (g)

[,10] 	 LS 	 numeric 	 Litter size

[,11] 	 GL 	 numeric 	 Gestation length (days)

[,12] 	 BW 	 numeric 	 Birth weigth (g)

[,13] 	 WA 	 numeric 	 Weaning age (days)

[,14] 	 AI 	 numeric 	 Age of independance (days)

[,15] 	 LY 	 numeric 	 Longevity (months)

[,16] 	 AM 	 numeric 	 Age of sexual maturity (days)

[,17] 	 IB 	 numeric 	 Inter-birth interval (months)

Source

Gittleman, J. L. (1986) Carnivore life history patterns: allometric, phylogenetic and ecological associations. American Naturalist, 127: 744–771.

Examples

data(carnivora)
## Fig. 1 in Gittleman (1986):
plot(carnivora$BW ~ carnivora$FW, pch = (1:8)[carnivora$Family], log = "xy",
     xlab = "Female body weight (kg)", ylab = "Birth weigth (g)",
     ylim = c(1, 2000))
legend("bottomright", legend = levels(carnivora$Family), pch = 1:8)
plot(carnivora$BW ~ carnivora$FB, pch = (1:8)[carnivora$Family], log = "xy",
     xlab = "Female brain weight (g)", ylab = "Birth weigth (g)",
     ylim = c(1, 2000))
legend("bottomright", legend = levels(carnivora$Family), pch = 1:8)

[Package ape version 5.2 Index]