checkLabel {ape}	R Documentation

Checking Labels

Description

Checking and correcting character strings, particularly before writing a Newick tree.

Usage

checkLabel(x)

Arguments

x
a vector of mode character.

Details

This function deletes the leading and trailing spaces (including tabulations, new lines, and left or right parentheses at the beginning or end of the strings), substitutes the spaces inside the strings by underscores, and substitutes commas, colons, semicolons, and parentheses inside the strings by dashes.

Value

a vector of mode character.

Author(s)

Emmanuel Paradis

See Also

makeLabel, makeNodeLabel, mixedFontLabel, stripLabel, updateLabel

Examples


checkLabel(" Homo sapiens\t(Primates; World)   ")

[Package ape version 5.5 Index]