checkValidPhylo {ape}	R Documentation

Check the Structure of a "phylo" Object

Description

This function takes as single argument an object (phy), checks its elements, and prints a diagnostic. All problems are printed with a label: FATAL (will likely cause an error or a crash) or MODERATE (may cause some problems).

This function is mainly intended for developers creating "phylo" objects from scratch.

Usage

checkValidPhylo(phy)

Arguments

phy
an object of class "phylo".

Value

NULL.

Author(s)

Emmanuel Paradis

Examples

tr <- rtree(3)
checkValidPhylo(tr)
tr$edge[1] <- 0
checkValidPhylo(tr)

[Package ape version 5.2 Index]