coalescent.intervals {ape}	R Documentation

Coalescent Intervals

Description

This function extracts or generates information about coalescent intervals (number of lineages, interval lengths, interval count, total depth) from a phylogenetic tree or a list of internode distances. The input tree needs to be ultra-metric (i.e. clock-like).

Usage

coalescent.intervals(x)

Arguments

x
either an ultra-metric phylogenetic tree (i.e. an object of class "phylo") or, alternatively, a vector of interval lengths.

Value

An object of class "coalescentIntervals" with the following entries:

lineages
A vector with the number of lineages at the start of each coalescent interval.

interval.length
A vector with the length of each coalescent interval.

interval.count
The total number of coalescent intervals.

total.depth
The sum of the lengths of all coalescent intervals.

Author(s)

Korbinian Strimmer

See Also

branching.times, collapsed.intervals, read.tree.

Examples

data("hivtree.newick") # example tree in NH format
tree.hiv <- read.tree(text = hivtree.newick) # load tree
ci <- coalescent.intervals(tree.hiv) # from tree
ci
data("hivtree.table") # same tree, but in table format
ci <- coalescent.intervals(hivtree.table$size) # from vector of interval lengths
ci

[Package ape version 5.2 Index]