cophenetic.phylo {ape}	R Documentation

Pairwise Distances from a Phylogenetic Tree

Description

cophenetic.phylo computes the pairwise distances between the pairs of tips from a phylogenetic tree using its branch lengths.

dist.nodes does the same but between all nodes, internal and terminal, of the tree.

Usage

## S3 method for class 'phylo'
cophenetic(x)
dist.nodes(x)

Arguments

x
an object of class "phylo".

Value

a numeric matrix with colnames and rownames set to the names of the tips (as given by the element tip.label of the argument phy), or, in the case of dist.nodes, the numbers of the tips and the nodes (as given by the element edge).

Author(s)

Emmanuel Paradis

See Also

read.tree to read tree files in Newick format, cophenetic for the generic function

[Package ape version 5.2 Index]