cophyloplot {ape}	R Documentation

Plots two phylogenetic trees face to face with links between the tips.

Description

This function plots two trees face to face with the links if specified. It is possible to rotate the branches of each tree around the nodes by clicking.

Usage

cophyloplot(x, y, assoc = NULL, use.edge.length = FALSE, space = 0,
       length.line = 1, gap = 2, type = "phylogram", rotate = FALSE,
       col = par("fg"), lwd = par("lwd"), lty = par("lty"),
       show.tip.label = TRUE, font = 3, ...)

Arguments

x, y
two objects of class "phylo".

assoc
a matrix with 2 columns specifying the associations between the tips. If NULL, no links will be drawn.

use.edge.length
a logical indicating whether the branch lengths should be used to plot the trees; default is FALSE.

space
a positive value that specifies the distance between the two trees.

length.line
a positive value that specifies the length of the horizontal line associated to each taxa. Default is 1.

gap
a value specifying the distance between the tips of the phylogeny and the lines.

type
a character string specifying the type of phylogeny to be drawn; it must be one of "phylogram" (the default) or "cladogram".

rotate
a logical indicating whether the nodes of the phylogeny can be rotated by clicking. Default is FALSE.

col
a character vector indicating the color to be used for the links; recycled as necessary.

lwd
id. for the width.

lty
id. for the line type.

show.tip.label
a logical indicating whether to show the tip labels on the phylogeny (defaults to 'TRUE', i.e. the labels are shown).

font
an integer specifying the type of font for the labels: 1 (plain text), 2 (bold), 3 (italic, the default), or 4 (bold italic).

...
(unused)

Details

The aim of this function is to plot simultaneously two phylogenetic trees with associated taxa. The two trees do not necessarily have the same number of tips and more than one tip in one phylogeny can be associated with a tip in the other.

The association matrix used to draw the links has to be a matrix with two columns containing the names of the tips. One line in the matrix represents one link on the plot. The first column of the matrix has to contain tip labels of the first tree (phy1) and the second column of the matrix, tip labels of the second tree (phy2). There is no limit (low or high) for the number of lines in the matrix. A matrix with two colums and one line will give a plot with one link.

Arguments gap, length.line and space have to be changed to get a nice plot of the two phylogenies. Note that the function takes into account the length of the character strings corresponding to the names at the tips, so that the lines do not overwrite those names.

The rotate argument can be used to transform both phylogenies in order to get the more readable plot (typically by decreasing the number of crossing lines). This can be done by clicking on the nodes. The escape button or right click take back to the console.

Author(s)

Damien de Vienne damien.de-vienne@u-psud.fr

See Also

plot.phylo, rotate, rotateConstr

Examples

#two random trees
tree1 <- rtree(40)
tree2 <- rtree(20)

#creation of the association matrix:
association <- cbind(tree2$tip.label, tree2$tip.label)

cophyloplot(tree1, tree2, assoc = association,
            length.line = 4, space = 28, gap = 3)

#plot with rotations
## Not run: 
cophyloplot(tree1, tree2, assoc=association, length.line=4, space=28, gap=3, rotate=TRUE)

## End(Not run)

[Package ape version 5.2 Index]