corBlomberg {ape}	R Documentation

Blomberg et al.'s Correlation Structure

Description

The “ACDC” (accelerated/decelerated) model assumes that continuous traits evolve under a Brownian motion model which rates accelerates (if g < 1) or decelerates (if g > 1) through time. If g = 1, then the model reduces to a Brownian motion model.

Usage

corBlomberg(value, phy, form = ~1, fixed = FALSE)
## S3 method for class 'corBlomberg'
corMatrix(object, covariate = getCovariate(object),
                   corr = TRUE, ...)
## S3 method for class 'corBlomberg'
coef(object, unconstrained = TRUE, ...)

Arguments

value
the (initial) value of the parameter g.

phy
an object of class "phylo".

form
(ignored).

fixed
a logical specifying whether gls should estimate gamma (the default) or keep it fixed.

object
an (initialized) object of class "corBlomberg".

covariate
(ignored).

corr
a logical value specifying whether to return the correlation matrix (the default) or the variance-covariance matrix.

unconstrained
a logical value. If TRUE (the default), the coefficients are returned in unconstrained form (the same used in the optimization algorithm). If FALSE the coefficients are returned in “natural”, possibly constrained, form.

...
further arguments passed to or from other methods.

Value

an object of class "corBlomberg", the coefficients from an object of this class, or the correlation matrix of an initialized object of this class. In most situations, only corBlomberg will be called by the user.

Author(s)

Emmanuel Paradis

References

Blomberg, S. P., Garland, Jr, T., and Ives, A. R. (2003) Testing for phylogenetic signal in comparative data: behavioral traits are more labile. Evolution, 57, 717–745.

[Package ape version 5.2 Index]