corBrownian {ape}	R Documentation

Brownian Correlation Structure

Description

Expected covariance under a Brownian model (Felsenstein 1985, Martins and Hansen 1997)

Vij = gamma . ta

where ta is the distance on the phylogeny between the root and the most recent common ancestor of taxa i and j and gamma is a constant.

Usage

corBrownian(value=1, phy, form=~1)
## S3 method for class 'corBrownian'
coef(object, unconstrained = TRUE, ...)
## S3 method for class 'corBrownian'
corMatrix(object, covariate = getCovariate(object), corr = TRUE, ...)

Arguments

value
The gamma parameter (default to 1)

phy
An object of class phylo representing the phylogeny (with branch lengths) to consider

object
An (initialized) object of class corBrownian

corr
a logical value. If 'TRUE' the function returns the correlation matrix, otherwise it returns the variance/covariance matrix.

form
ignored for now.

covariate
ignored for now.

unconstrained
a logical value. If 'TRUE' the coefficients are returned in unconstrained form (the same used in the optimization algorithm). If 'FALSE' the coefficients are returned in "natural", possibly constrained, form. Defaults to 'TRUE'

...
some methods for these generics require additional arguments. None are used in these methods.

Value

An object of class corBrownian, or the coefficient from an object of this class (actually sends numeric(0)), or the correlation matrix of an initialized object of this class.

Author(s)

Julien Dutheil julien.dutheil@univ-montp2.fr

References

Felsenstein, J. (1985) Phylogenies and the comparative method. American Naturalist, 125, 1–15.

Martins, E. P. and Hansen, T. F. (1997) Phylogenies and the comparative method: a general approach to incorporating phylogenetic information into the analysis of interspecific data. American Naturalist, 149, 646–667.

See Also

corClasses

[Package ape version 5.2 Index]