corGrafen {ape}	R Documentation

Grafen's (1989) Correlation Structure

Description

Grafen's (1989) covariance structure. Branch lengths are computed using Grafen's method (see compute.brlen). The covariance matrice is then the traditional variance-covariance matrix for a phylogeny.

Usage

corGrafen(value, phy, form=~1, fixed = FALSE)
## S3 method for class 'corGrafen'
coef(object, unconstrained = TRUE, ...)
## S3 method for class 'corGrafen'
corMatrix(object,
                  covariate = getCovariate(object), corr = TRUE, ...)

Arguments

value
The rho parameter

phy
An object of class phylo representing the phylogeny (branch lengths are ignored) to consider

object
An (initialized) object of class corGrafen

corr
a logical value. If 'TRUE' the function returns the correlation matrix, otherwise it returns the variance/covariance matrix.

fixed
an optional logical value indicating whether the coefficients should be allowed to vary in the optimization, or kept fixed at their initial value. Defaults to 'FALSE', in which case the coefficients are allowed to vary.

form
ignored for now.

covariate
ignored for now.

unconstrained
a logical value. If 'TRUE' the coefficients are returned in unconstrained form (the same used in the optimization algorithm). If 'FALSE' the coefficients are returned in "natural", possibly constrained, form. Defaults to 'TRUE'

...
some methods for these generics require additional arguments. None are used in these methods.

Value

An object of class corGrafen or the rho coefficient from an object of this class or the correlation matrix of an initialized object of this class.

Author(s)

Julien Dutheil julien.dutheil@univ-montp2.fr

References

Grafen, A. (1989) The phylogenetic regression. Philosophical Transactions of the Royal society of London. Series B. Biological Sciences, 326, 119–157.

See Also

corClasses, compute.brlen, vcv.phylo.

[Package ape version 5.2 Index]