corMartins {ape}	R Documentation

Martins's (1997) Correlation Structure

Description

Martins and Hansen's (1997) covariance structure:

Vij = gamma . exp(-alpha . tij)

where tij is the phylogenetic distance between taxa i and j and gamma is a constant.

Usage

corMartins(value, phy, form = ~1, fixed = FALSE)
## S3 method for class 'corMartins'
coef(object, unconstrained = TRUE, ...)
## S3 method for class 'corMartins'
corMatrix(object,
		covariate = getCovariate(object), corr = TRUE, ...)

Arguments

value
The alpha parameter

phy
An object of class phylo representing the phylogeny (with branch lengths) to consider

object
An (initialized) object of class corMartins

corr
a logical value. If 'TRUE' the function returns the correlation matrix, otherwise it returns the variance/covariance matrix.

fixed
an optional logical value indicating whether the coefficients should be allowed to vary in the optimization, ok kept fixed at their initial value. Defaults to 'FALSE', in which case the coefficients are allowed to vary.

form
ignored for now.

covariate
ignored for now.

unconstrained
a logical value. If 'TRUE' the coefficients are returned in unconstrained form (the same used in the optimization algorithm). If 'FALSE' the coefficients are returned in "natural", possibly constrained, form. Defaults to 'TRUE'

...
some methods for these generics require additional arguments. None are used in these methods.

Value

An object of class corMartins or the alpha coefficient from an object of this class or the correlation matrix of an initialized object of this class.

Author(s)

Julien Dutheil julien.dutheil@univ-montp2.fr

References

Martins, E. P. and Hansen, T. F. (1997) Phylogenies and the comparative method: a general approach to incorporating phylogenetic information into the analysis of interspecific data. American Naturalist, 149, 646–667.

See Also

corClasses.

[Package ape version 5.2 Index]