correlogram.formula {ape}	R Documentation

Phylogenetic Correlogram

Description

This function computes a correlogram from taxonomic levels.

Usage

  correlogram.formula(formula, data = NULL, use = "all.obs")

Arguments

formula
a formula of the type y1+..+yn ~ g1/../gn, where the y's are the data to analyse and the g's are the taxonomic levels.

data
a data frame containing the variables specified in the formula. If NULL, the variables are sought in the user's workspace.

use
a character string specifying how to handle missing values (i.e., NA). This must be one of "all.obs", "complete.obs", or "pairwise.complete.obs", or any unambiguous abbrevation of these. In the first case, the presence of missing values produces an error. In the second case, all rows with missing values will be removed before computation. In the last case, missing values are removed on a case-by-case basis.

Details

See the vignette in R: vignette("MoranI").

Value

An object of class correlogram which is a data frame with three columns:

obs
the computed Moran's I

p.values
the corresponding P-values

labels
the names of each level

or an object of class correlogramList containing a list of objects of class correlogram if several variables are given as response in formula.

Author(s)

Julien Dutheil julien.dutheil@univ-montp2.fr and Emmanuel Paradis

See Also

plot.correlogram, Moran.I

Examples

data(carnivora)
### Using the formula interface:
co <- correlogram.formula(SW ~ Order/SuperFamily/Family/Genus,
      data=carnivora)
co
plot(co)
### Several correlograms on the same plot:
cos <- correlogram.formula(SW + FW ~ Order/SuperFamily/Family/Genus,
      data=carnivora)
cos
plot(cos)

[Package ape version 5.2 Index]