delta.plot {ape}	R Documentation

Delta Plots

Description

This function makes a delta plot following Holland et al. (2002).

Usage

delta.plot(X, k = 20, plot = TRUE, which = 1:2)

Arguments

X
a distance matrix, may be an object of class “dist”.

k
an integer giving the number of intervals in the plot.

plot
a logical specifying whether to draw the delta plot (the default).

which
a numeric vector indicating which plots are done; 1: the histogram of the delta_q values, 2: the plot of the individual delta.bar values. By default, both plots are done.

Details

See Holland et al. (2002) for details and interpretation.

The computing time of this function is proportional to the fourth power of the number of observations (O(n^4)), so calculations may be very long with only a slight increase in sample size.

Value

This function returns invisibly a named list with two components:

countsthe counts for the histogram of delta_q values

delta.barthe mean delta value for each observation

Author(s)

Emmanuel Paradis

References

Holland, B. R., Huber, K. T., Dress, A. and Moulton, V. (2002) Delta plots: a tool for analyzing phylogenetic distance data. Molecular Biology and Evolution, 12, 2051–2059.

See Also

dist.dna

Examples

data(woodmouse)
d <- dist.dna(woodmouse)
delta.plot(d)
layout(1)
delta.plot(d, 40, which = 1)

[Package ape version 5.2 Index]