dnds {ape}	R Documentation

dN/dS Ratio

Description

This function computes the pairwise ratios dN/dS for a set of aligned DNA sequences using Li's (1993) method.

Usage


dnds(x, code = 1, codonstart = 1, quiet = FALSE)

Arguments

x
an object of class "DNAbin" (matrix or list) with the aligned sequences.

code
an integer value giving the genetic code to be used. Currently only 1 (standard code) and 2 (vertebrate mitochondrial code) are supported.

codonstart
an integer giving where to start the translation. This should be 1, 2, or 3, but larger values are accepted and have for effect to start the translation further within the sequence.

quiet
single logical value: whether to indicate progress of calculations.

Value

an object of class "dist".

Author(s)

Emmanuel Paradis

References

Li, W.-H. (1993) Unbiased estimation of the rates of synonymous and nonsynonymous substitution. Journal of Molecular Evolution, 36, 96–99.

See Also

AAbin, trans, alview

Examples


data(woodmouse)
res <- dnds(woodmouse, quiet = TRUE) # NOT correct
res2 <- dnds(woodmouse, code = 2, quiet = TRUE) # using the correct code
identical(res, res2)

[Package ape version 5.5 Index]