FastME {ape}	R Documentation

Tree Estimation Based on the Minimum Evolution Algorithm

Description

The two FastME functions (balanced and OLS) perform the minimum evolution algorithm of Desper and Gascuel (2002).

Usage

  fastme.bal(X, nni = TRUE, spr = TRUE, tbr = TRUE)
  fastme.ols(X, nni = TRUE)

Arguments

X
a distance matrix; may be an object of class "dist".

nni
a boolean value; TRUE to do NNIs (default).

spr
ditto for SPRs.

tbr
ditto for TBRs.

Value

an object of class "phylo".

Author(s)

original C code by Richard Desper; adapted and ported to R by Vincent Lefort vincent.lefort@lirmm.fr

References

Desper, R. and Gascuel, O. (2002) Fast and accurate phylogeny reconstruction algorithms based on the minimum-evolution principle. Journal of Computational Biology, 9(5), 687–705.

See Also

nj, bionj, write.tree, read.tree, dist.dna

Examples

### From Saitou and Nei (1987, Table 1):
x <- c(7, 8, 11, 13, 16, 13, 17, 5, 8, 10, 13,
       10, 14, 5, 7, 10, 7, 11, 8, 11, 8, 12,
       5, 6, 10, 9, 13, 8)
M <- matrix(0, 8, 8)
M[lower.tri(M)] <- x
M <- t(M)
M[lower.tri(M)] <- x
dimnames(M) <- list(1:8, 1:8)
tr <- fastme.bal(M)
plot(tr, "u")
### a less theoretical example
data(woodmouse)
trw <- fastme.bal(dist.dna(woodmouse))
plot(trw)

[Package ape version 5.2 Index]