is.compatible {ape}	R Documentation

Check Compatibility of Splits

Description

is.compatible is a generic function with a method for the class "bitsplits". It checks whether a set of splits is compatible using the arecompatible function.

Usage

is.compatible(obj)
## S3 method for class 'bitsplits'
is.compatible(obj)
arecompatible(x, y, n)

Arguments

obj
an object of class "bitsplits".

x, y
a vector of mode raw.

n
the number of taxa in the splits.

Value

TRUE if the splits are compatible, FALSE otherwise.

Author(s)

Andrei Popescu niteloserpopescu@gmail.com

See Also

as.bitsplits

[Package ape version 5.2 Index]