is.monophyletic {ape}	R Documentation

Is Group Monophyletic

Description

This function tests whether a list of tip labels is monophyletic on a given tree.

Usage

is.monophyletic(phy, tips, reroot = !is.rooted(phy), plot = FALSE, ...)

Arguments

phy
a phylogenetic tree description of class "phylo".

tips
a vector of mode numeric or character specifying the tips to be tested.

reroot
a logical. If FALSE, then the input tree is not unrooted before the test.

plot
a logical. If TRUE, then the tree is plotted with the specified group tips highlighted.

...
further arguments passed to plot.

Details

If phy is rooted, the test is done on the rooted tree, otherwise the tree is first unrooted, then arbitrarily rerooted, in order to be independent on the current position of the root. That is, the test asks if tips could be monophyletic given any favourably rooting of phy.

If phy is unrooted the test is done on an unrooted tree, unless reroot = FALSE is specified.

If tip labels in the list tips are given as characters, they need to be spelled as in the object phy.

Value

TRUE or FALSE.

Author(s)

Johan Nylander jnylander@users.sourceforge.net

See Also

which.edge, drop.tip, mrca.

Examples

    ## Test one monophyletic and one paraphyletic group on the bird.orders tree
    ## Not run: data("bird.orders")
    ## Not run: is.monophyletic(phy = bird.orders, tips = c("Ciconiiformes", "Gruiformes"))
    ## Not run: is.monophyletic(bird.orders, c("Passeriformes", "Ciconiiformes", "Gruiformes"))
    

[Package ape version 5.2 Index]