is.ultrametric {ape}	R Documentation

Test if a Tree is Ultrametric

Description

This function tests whether a tree is ultrametric using the distances from each tip to the root.

Usage

is.ultrametric(phy, ...)
## S3 method for class 'phylo'
is.ultrametric(phy, tol = .Machine$double.eps^0.5, option = 1, ...)
## S3 method for class 'multiPhylo'
is.ultrametric(phy, tol = .Machine$double.eps^0.5, option = 1, ...)

Arguments

phy
an object of class "phylo" or "multiPhylo".

tol
a numeric >= 0, variation below this value are considered non-significant.

option
an integer (1 or 2; see details).

...
arguments passed among methods.

Details

The test is based on the distances from each tip to the root and a criterion: if option = 1, the criterion is the scaled range ((max - min/max)), if option = 2, the variance is used (this was the method used until ape 3.5). The default criterion is invariant to linear changes of the branch lengths.

Value

a logical vector.

Author(s)

Emmanuel Paradis

See Also

is.binary, .Machine

Examples

is.ultrametric(rtree(10))
is.ultrametric(rcoal(10))

[Package ape version 5.2 Index]