kronoviz {ape}	R Documentation

Plot Multiple Chronograms on the Same Scale

Description

The main argument is a list of (rooted) trees which are plotted on the same scale.

Usage

kronoviz(x, layout = length(x), horiz = TRUE, ...)

Arguments

x
a list of (rooted) trees of class "phylo".

layout
an integer giving the number of trees plotted simultaneously; by default all.

horiz
a logical specifying whether the trees should be plotted rightwards (the default) or upwards.

...
further arguments passed to plot.phylo.

Details

The size of the individual plots is proportional to the size of the trees.

Value

NULL

Author(s)

Emmanuel Paradis

See Also

plot.phylo

Examples

TR <- replicate(10, rcoal(sample(11:20, size = 1)), simplify = FALSE)
kronoviz(TR)
kronoviz(TR, horiz = FALSE, type = "c", show.tip.label = FALSE)

[Package ape version 5.2 Index]