label2table {ape}	R Documentation

Label Management

Description

These functions work on a vector of character strings storing bi- or trinomial species names, typically “Genus_species_subspecies”.

Usage

label2table(x, sep = NULL, as.is = FALSE)
stripLabel(x, species = FALSE, subsp = TRUE, sep = NULL)
abbreviateGenus(x, genus = TRUE, species = FALSE, sep = NULL)

Arguments

x
a vector of mode character.

sep
the separator (a single character) between the taxonomic levels (see details).

as.is
a logical specifying whether to convert characters into factors (like in read.table).

species, subsp, genus
a logical specifying whether the taxonomic level is concerned by the operation.

Details

label2table returns a data frame with three columns named “genus”, “species”, and “subspecies” (with NA if the level is missing).

stripLabel deletes the subspecies names from the input. If species = TRUE, the species names are also removed, thus returning only the genus names.

abbreviateGenus abbreviates the genus names keeping only the first letter. If species = TRUE, the species names are abbreviated.

By default, these functions try to guess what is the separator between the genus, species and/or subspecies names. If an underscore is present in the input, then this character is assumed to be the separator; otherwise, a space. If this does not work, you can specify sep to its appropriate value.

Value

A vector of mode character or a data frame.

Author(s)

Emmanuel Paradis

See Also

makeLabel, makeNodeLabel, mixedFontLabel, updateLabel

Examples

x <- c("Panthera_leo", "Panthera_pardus", "Panthera_onca", "Panthera_uncia",
       "Panthera_tigris_altaica", "Panthera_tigris_amoyensis")
label2table(x)
stripLabel(x)
stripLabel(x, TRUE)
abbreviateGenus(x)
abbreviateGenus(x, species = TRUE)
abbreviateGenus(x, genus = FALSE, species = TRUE)

[Package ape version 5.2 Index]