ladderize {ape}	R Documentation

Ladderize a Tree

Description

This function reorganizes the internal structure of the tree to get the ladderized effect when plotted.

Usage

ladderize(phy, right = TRUE)

Arguments

phy
an object of class "phylo".

right
a logical specifying whether the smallest clade is on the right-hand side (when the tree is plotted upwards), or the opposite (if FALSE).

Author(s)

Emmanuel Paradis

See Also

plot.phylo, reorder.phylo

Examples

tr <- rcoal(50)
layout(matrix(1:4, 2, 2))
plot(tr, main = "normal")
plot(ladderize(tr), main = "right-ladderized")
plot(ladderize(tr, FALSE), main = "left-ladderized")
layout(matrix(1, 1))

[Package ape version 5.2 Index]