latag2n {ape}	R Documentation

Leading and Trailing Alignment Gaps to N

Description

Substitutes leading and trailing alignment gaps in aligned sequences into N (i.e., A, C, G, or T). The gaps in the middle of the sequences are left unchanged.

Usage


latag2n(x)

Arguments

x
an object of class "DNAbin" with the aligned sequences.

Details

This function is called by others in ape and in pegas. It is documented here in case it needs to be called by other packages.

Value

an object of class "DNAbin".

Author(s)

Emmanuel Paradis

See Also

DNAbin

Examples


x <- as.DNAbin(matrix(c("-", "A", "G", "-", "T", "C"), 2, 3))
y <- latag2n(x)
alview(x)
alview(y)

[Package ape version 5.5 Index]