matexpo {ape}	R Documentation

Matrix Exponential

Description

This function computes the exponential of a square matrix using a spectral decomposition.

Usage

matexpo(x)

Arguments

x
a square matrix of mode numeric.

Value

a numeric matrix of the same dimensions than ‘x’.

Author(s)

Emmanuel Paradis

Examples

### a simple rate matrix:
m <- matrix(0.1, 4, 4)
diag(m) <- -0.3
### towards equilibrium:
for (t in c(1, 5, 10, 50)) print(matexpo(m*t))

[Package ape version 5.2 Index]