mrca {ape}	R Documentation

Find Most Recent Common Ancestors Between Pairs

Description

mrca returns for each pair of tips (and nodes) its most recent common ancestor (MRCA).

getMRCA returns the MRCA of two or more tips.

Usage

mrca(phy, full = FALSE)
getMRCA(phy, tip)

Arguments

phy
an object of class "phylo".

full
a logical indicating whether to return the MRCAs among all tips and nodes (if TRUE); the default is to return only the MRCAs among tips.

tip
a vector of mode numeric or character specifying the tips; can also be node numbers.

Details

For mrca, the diagonal is set to the number of the tips (and nodes if full = TRUE). If full = FALSE, the colnames and rownames are set with the tip labels of the tree; otherwise the numbers are given as names.

For getMRCA, if tip is of length one or zero then NULL is returned.

Value

a matrix of mode numeric (mrca) or a single numeric value (getMRCA).

Author(s)

Emmanuel Paradis, Klaus Schliep, Joseph W. Brown

[Package ape version 5.2 Index]