multi2di {ape}	R Documentation

Collapse and Resolve Multichotomies

Description

These two functions collapse or resolve multichotomies in phylogenetic trees.

Usage

multi2di(phy, ...)
## S3 method for class 'phylo'
multi2di(phy, random = TRUE, ...)
## S3 method for class 'multiPhylo'
multi2di(phy, random = TRUE, ...)
di2multi(phy, ...)
## S3 method for class 'phylo'
di2multi(phy, tol = 1e-08, ...)
## S3 method for class 'multiPhylo'
di2multi(phy, tol = 1e-08, ...)

Arguments

phy
an object of class "phylo" or "multiPhylo".

random
a logical value specifying whether to resolve the multichotomies randomly (the default) or in the order they appear in the tree (if random = FALSE).

tol
a numeric value giving the tolerance to consider a branch length significantly greater than zero.

...
arguments passed among methods.

Details

multi2di transforms all multichotomies into a series of dichotomies with one (or several) branch(es) of length zero.

di2multi deletes all branches smaller than tol and collapses the corresponding dichotomies into a multichotomy.

Value

an object of the same class than the input.

Author(s)

Emmanuel Paradis

See Also

is.binary

Examples

data(bird.families)
is.binary(bird.families)
is.binary(multi2di(bird.families))
all.equal(di2multi(multi2di(bird.families)), bird.families)
### To see the results of randomly resolving a trichotomy:
tr <- read.tree(text = "(a:1,b:1,c:1);")
layout(matrix(1:4, 2, 2))
for (i in 1:4)
  plot(multi2di(tr), use.edge.length = FALSE, cex = 1.5)
layout(1)

[Package ape version 5.2 Index]