mvr {ape}	R Documentation

Minimum Variance Reduction

Description

Phylogenetic tree construction based on the minimum variance reduction.

Usage

mvr(X, V)
mvrs(X, V, fs = 15)

Arguments

X
a distance matrix.

V
a variance matrix.

fs
agglomeration criterion parameter: it is coerced as an integer and must at least equal to one.

Details

The MVR method can be seen as a version of BIONJ which is not restricted to the Poison model of variance (Gascuel 2000).

Value

an object of class "phylo".

Author(s)

Andrei Popescu niteloserpopescu@gmail.com

References

Criscuolo, A. and Gascuel, O. (2008). Fast NJ-like algorithms to deal with incomplete distance matrices. BMC Bioinformatics, 9.

Gascuel, O. (2000). Data model and classification by trees: the minimum variance reduction (MVR) method. Journal of Classification, 17, 67–99.

See Also

bionj, fastme, njs, SDM

Examples

data(woodmouse)
rt <- dist.dna(woodmouse, variance = TRUE)
v <- attr(rt, "variance")
tr <- mvr(rt, v)
plot(tr, "u")

[Package ape version 5.2 Index]