node.depth {ape}	R Documentation

Depth and Heights of Nodes and Tips

Description

These functions return the depths or heights of nodes and tips.

Usage

node.depth(phy, method = 1)
node.depth.edgelength(phy)
node.height(phy, clado.style = FALSE)

Arguments

phy
an object of class "phylo".

method
an integer value (1 or 2); 1: the node depths are proportional to the number of tips descending from each node, 2: they are evenly spaced.

clado.style
a logical value; if TRUE, the node heights are calculated for a cladogram.

Details

node.depth computes the depth of a node depending on the value of method (see the option node.depth in plot.phylo). The value of 1 is given to the tips.

node.depth.edgelength does the same but using branch lengths.

node.height computes the heights of nodes and tips as plotted by a phylogram or a cladogram.

Value

A numeric vector indexed with the node numbers of the matrix ‘edge’ of phy.

Author(s)

Emmanuel Paradis

See Also

plot.phylo

[Package ape version 5.2 Index]