plot.correlogram {ape}	R Documentation

Plot a Correlogram

Description

These functions plot correlagrams previously computed with correlogram.formula.

Usage

  ## S3 method for class 'correlogram'
plot(x, legend = TRUE, test.level = 0.05,
                col = c("grey", "red"), type = "b", xlab = "",
                ylab = "Moran's I", pch = 21, cex = 2, ...)
  ## S3 method for class 'correlogramList'
plot(x, lattice = TRUE, legend = TRUE,
                test.level = 0.05, col = c("grey", "red"),
                xlab = "", ylab = "Moran's I",
                type = "b", pch = 21, cex = 2, ...)

Arguments

x
an object of class "correlogram" or of class "correlogramList" (both produced by correlogram.formula).

legend
should a legend be added on the plot?

test.level
the level used to discriminate the plotting symbols with colours considering the P-values.

col
two colours for the plotting symbols: the first one is used if the P-value is greater than or equal to test.level, the second one otherwise.

type
the type of plot to produce (see plot for possible choices).

xlab
an optional character string for the label on the x-axis (none by default).

ylab
the default label on the y-axis.

pch
the type of plotting symbol.

cex
the default size for the plotting symbols.

lattice
when plotting several correlograms, should they be plotted in trellis-style with lattice (the default), or together on the same plot?

...
other parameters passed to the plot or lines function.

Details

When plotting several correlograms with lattice, some options have no effect: legend, type, and pch (pch=19 is always used in this situation).

When using pch between 1 and 20 (i.e., non-filled symbols, the colours specified in col are also used for the lines joining the points. To keep black lines, it is better to leave pch between 21 and 25.

Author(s)

Emmanuel Paradis

See Also

correlogram.formula, Moran.I

[Package ape version 5.2 Index]