plotTreeTime {ape}	R Documentation

Plot Tree With Time Axis

Description

This function plots a non-ultrametric tree where the tips are not contemporary together with their dates on the x-axis.

Usage

plotTreeTime(phy, tip.dates, show.tip.label = FALSE, y.lim = NULL,
             color = TRUE, ...)

Arguments

phy
an object of class "phylo".

tip.dates
a vector of the same length than the number of tips in phy (see details).

show.tip.label
a logical value; see plot.phylo.

y.lim
by default, one fifth of the plot is left below the tree; use this option to change this behaviour.

color
a logical value specifying whether to use colors for the lines linking the tips to the time axis. If FALSE, a grey scale is used.

...
other arguments to be passed to plot.phylo.

Details

The vector tip.dates may be numeric or of class “Date”. In either case, the time axis is set accordingly. The length of this vector must be equal to the number of tips of the tree: the dates are matched to the tips numbers. Missing values are allowed.

Value

NULL

Author(s)

Emmanuel Paradis

See Also

plot.phylo, estimate.dates

Examples

dates <- as.Date(.leap.seconds)
tr <- rtree(length(dates))
plotTreeTime(tr, dates)

## handling NA's:
dates[11:26] <- NA
plotTreeTime(tr, dates)

## dates can be on an arbitrary scale, e.g., [-1, 1]:
plotTreeTime(tr, runif(Ntip(tr), -1, 1))

[Package ape version 5.2 Index]