print.phylo {ape}	R Documentation

Compact Display of a Phylogeny

Description

These functions prints a compact summary of a phylogeny, or a list of phylogenies, on the console.

Usage

## S3 method for class 'phylo'
print(x, printlen = 6 ,...)
## S3 method for class 'multiPhylo'
print(x, details = FALSE ,...)
## S3 method for class 'multiPhylo'
str(object, ...)

Arguments

x
an object of class "phylo" or "multiPhylo".

object
an object of class "multiPhylo".

printlen
the number of labels to print (6 by default).

details
a logical indicating whether to print information on all trees.

...
further arguments passed to or from other methods.

Value

NULL.

Author(s)

Ben Bolker and Emmanuel Paradis

See Also

read.tree, summary.phylo, print for the generic R function

Examples

x <- rtree(10)
print(x)
print(x, printlen = 10)
x <- rmtree(2, 10)
print(x)
print(x, TRUE)
str(x)

[Package ape version 5.2 Index]