read.caic {ape}	R Documentation

Read Tree File in CAIC Format

Description

This function reads one tree from a CAIC file. A second file containing branch lengths values may also be passed (experimental).

Usage

read.caic(file, brlen = NULL, skip = 0, comment.char = "#", ...)

Arguments

file
a file name specified by either a variable of mode character, or a double-quoted string.

brlen
a file name for the branch lengths file.

skip
the number of lines of the input file to skip before beginning to read data (this is passed directly to scan()).

comment.char
a single character, the remaining of the line after this character is ignored (this is passed directly to scan()).

...
Further arguments to be passed to scan().

Details

Read a tree from a file in the format used by the CAIC and MacroCAIc program.

Value

an object of class "phylo".

Warning

The branch length support is still experimental and was not fully tested.

Author(s)

Julien Dutheil julien.dutheil@univ-montp2.fr

References

Purvis, A. and Rambaut, A. (1995) Comparative analysis by independent contrasts (CAIC): an Apple Macintosh application for analysing comparative data. CABIOS, 11 :241–251.

See Also

read.tree, read.nexus

Examples

### The same example than in read.tree, without branch lengths.
### An extract from Sibley and Ahlquist (1990)
cat("AAA","Strix_aluco","AAB","Asio_otus",
   "AB","Athene_noctua","B","Tyto_alba",
   file = "ex.tre", sep = "\n")
tree.owls <- read.caic("ex.tre")
plot(tree.owls)
tree.owls
unlink("ex.tre") # delete the file "ex.tre"

[Package ape version 5.2 Index]