richness.yule.test {ape}	R Documentation

Test of Diversification-Shift With the Yule Process

Description

This function performs a test of shift in diversification rate using probabilities from the Yule process.

Usage

richness.yule.test(x, t)

Arguments

x
a matrix or a data frame with at least two columns: the first one gives the number of species in clades with a trait supposed to increase or decrease diversification rate, and the second one the number of species in the sister-clades without the trait. Each row represents a pair of sister-clades.

t
a numeric vector giving the divergence times of each pair of clades in x.

Value

a data frame with the chi2, the number of degrees of freedom (= 1), and the P-value.

Author(s)

Emmanuel Paradis

References

Paradis, E. (2012) Shift in diversification in sister-clade comparisons: a more powerful test. Evolution, 66, 288–295.

See Also

slowinskiguyer.test, mcconwaysims.test, diversity.contrast.test

Examples

### see example(mcconwaysims.test)

[Package ape version 5.2 Index]