seg.sites {ape}	R Documentation

Find Segregating Sites in DNA Sequences

Description

This function gives the indices of segregating (polymorphic) sites in a sample of DNA sequences.

Usage

seg.sites(x)

Arguments

x
a matrix or a list which contains the DNA sequences.

Details

If the sequences are in a list, all the sequences must be of the same length. Ambiguous nucleotides are ignored.

Value

A numeric (integer) vector giving the indices of the segregating sites.

Author(s)

Emmanuel Paradis

See Also

base.freq, theta.s, nuc.div

Examples

data(woodmouse)
y <- seg.sites(woodmouse)
y
length(y)

[Package ape version 5.2 Index]