slowinskiguyer.test {ape}	R Documentation

Slowinski-Guyer Test of Homogeneous Diversification

Description

This function performs the Slowinski–Guyer test that a trait or variable does not increase diversification rate.

Usage

slowinskiguyer.test(x, detail = FALSE)

Arguments

x
a matrix or a data frame with at least two columns: the first one gives the number of species in clades with a trait supposed to increase diversification rate, and the second one the number of species in the corresponding sister-clade without the trait. Each row represents a pair of sister-clades.

detail
if TRUE, the individual P-values are appended.

Details

The Slowinski–Guyer test compares a series of sister-clades where one of the two is characterized by a trait supposed to increase diversification rate. The null hypothesis is that the trait does not affect diversification. If the trait decreased diversification rate, then the null hypothesis cannot be rejected.

The present function has mainly a historical interest. The Slowinski–Guyer test generally performs poorly: see Paradis (2012) alternatives and the functions cited below.

Value

a data frame with the chi2, the number of degrees of freedom, and the P-value. If detail = TRUE, a list is returned with the data frame and a vector of individual P-values for each pair of sister-clades.

Author(s)

Emmanuel Paradis

References

Paradis, E. (2012) Shift in diversification in sister-clade comparisons: a more powerful test. Evolution, 66, 288–295.

Slowinski, J. B. and Guyer, C. (1993) Testing whether certain traits have caused amplified diversification: an improved method based on a model of random speciation and extinction. American Naturalist, 142, 1019–1024.

See Also

balance, mcconwaysims.test, diversity.contrast.test, richness.yule.test, rc in geiger, shift.test in apTreeshape

Examples

### from Table 1 in Slowinski and Guyer(1993):
viviparous <- c(98, 8, 193, 36, 7, 128, 2, 3, 23, 70)
oviparous <- c(234, 17, 100, 4, 1, 12, 6, 1, 481, 11)
x <- data.frame(viviparous, oviparous)
slowinskiguyer.test(x, TRUE) # 'P ~ 0.32' in the paper
xalt <- x
xalt[3, 2] <- 1
slowinskiguyer.test(xalt)

[Package ape version 5.2 Index]