stree {ape}	R Documentation

Generates Systematic Regular Trees

Description

This function generates trees with regular shapes.

Usage

stree(n, type = "star", tip.label = NULL)

Arguments

n
an integer giving the number of tips in the tree.

type
a character string specifying the type of tree to generate; four choices are possible: "star", "balanced", "left", "right", or any unambiguous abbreviation of these.

tip.label
a character vector giving the tip labels; if not specified, the tips "t1", "t2", ..., are given.

Details

The types of trees generated are:

“star”a star (or comb) tree with a single internal node.

“balanced”a fully balanced dichotomous rooted tree; n must be a power of 2 (2, 4, 8, ...).

“left”a fully unbalanced rooted tree where the largest clade is on the left-hand side when the tree is plotted upwards.

“right”same than above but in the other direction.

Value

An object of class "phylo".

Author(s)

Emmanuel Paradis

See Also

compute.brlen, rtree

Examples

layout(matrix(1:4, 2, 2))
plot(stree(100))
plot(stree(128, "balanced"))
plot(stree(100, "left"))
plot(stree(100, "right"))

[Package ape version 5.2 Index]