treePop {ape}	R Documentation

Tree Popping

Description

Method for reconstructing phylogenetic trees from an object of class splits using tree popping.

Usage

treePop(obj)

Arguments

obj
an object of class "bitsplit".

Value

an object of class "phylo" which displays all the splits in the input object.

Author(s)

Andrei Popescu niteloserpopescu@gmail.com

[Package ape version 5.2 Index]