varcomp {ape}	R Documentation

Compute Variance Component Estimates

Description

Get variance component estimates from a fitted lme object.

Usage

varcomp(x, scale = FALSE, cum = FALSE)

Arguments

x
A fitted lme object

scale
Scale all variance so that they sum to 1

cum
Send cumulative variance components.

Details

Variance computations is done as in Venables and Ripley (2002).

Value

A named vector of class varcomp with estimated variance components.

Author(s)

Julien Dutheil julien.dutheil@univ-montp2.fr

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S (Fourth Edition). New York: Springer-Verlag.

See Also

lme

Examples

data(carnivora)
library(nlme)
m <- lme(log10(SW) ~ 1, random = ~ 1|Order/SuperFamily/Family/Genus, data=carnivora)
v <- varcomp(m, TRUE, TRUE)
plot(v)

[Package ape version 5.2 Index]