vcv {ape}	R Documentation

Phylogenetic Variance-covariance or Correlation Matrix

Description

This function computes the expected variances and covariances of a continuous trait assuming it evolves under a given model.

This is a generic function with methods for objects of class "phylo" and "corPhyl".

Usage

vcv(phy, ...)
## S3 method for class 'phylo'
vcv(phy, model = "Brownian", corr = FALSE, ...)
## S3 method for class 'corPhyl'
vcv(phy, corr = FALSE, ...)

Arguments

phy
an object of the correct class (see above).

model
a character giving the model used to compute the variances and covariances; only "Brownian" is available (for other models, a correlation structure may be used).

corr
a logical indicating whether the correlation matrix should be returned (TRUE); by default the variance-covariance matrix is returned (FALSE).

...
further arguments to be passed to or from other methods.

Value

a numeric matrix with the names of the tips as colnames and rownames.

Note

Do not confuse this function with vcov which computes the variance-covariance matrix among parameters of a fitted model object.

Author(s)

Emmanuel Paradis

References

Garland, T. Jr. and Ives, A. R. (2000) Using the past to predict the present: confidence intervals for regression equations in phylogenetic comparative methods. American Naturalist, 155, 346–364.

See Also

corBrownian, corMartins, corGrafen, corPagel, corBlomberg, vcv2phylo

Examples

tr <- rtree(5)
## all are the same:
vcv(tr)
vcv(corBrownian(1, tr))
vcv(corPagel(1, tr))

[Package ape version 5.2 Index]