vcv2phylo {ape}	R Documentation

Variance-Covariance Matrix to Tree

Description

This function transforms a variance-covariance matrix into a phylogenetic tree.

Usage

vcv2phylo(mat, tolerance = 1e-7)

Arguments

mat
a square symmetric (positive-definite) matrix.

tolerance
the numeric tolerance used to compare the branch lengths.

Details

The function tests if the matrix is symmetric and positive-definite (i.e., all its eigenvalues positive within the specified tolerance).

Value

an object of class "phylo".

Author(s)

Simon Blomberg

See Also

vcv, corPhyl

Examples

tr <- rtree(10)
V <- vcv(tr) # VCV matrix assuming Brownian motion
z <- vcv2phylo(V)
identical(tr, z) # FALSE
all.equal(tr, z) # TRUE

[Package ape version 5.2 Index]