where {ape}	R Documentation

Find Patterns in DNA Sequences

Description

This function finds patterns in a single or a set of DNA sequences.

Usage

where(x, pattern)

Arguments

x
an object of class "DNAbin".

pattern
a character string to be searched in x.

Details

If x is a vector, the function returns a single vector giving the position(s) where the pattern was found. If x is a matrix or a list, it returns a list with the positions of the pattern for each sequence.

Patterns may be overlapping. For instance, if pattern = "tata" and the sequence starts with ‘tatata’, then the output will be c(1, 3).

Value

a vector of integers or a list of such vectors.

Author(s)

Emmanuel Paradis

See Also

DNAbin, image.DNAbin

Examples

data(woodmouse)
where(woodmouse, "tata")

[Package ape version 5.2 Index]