which.edge {ape}	R Documentation

Identifies Edges of a Tree

Description

This function identifies the edges that belong to a group (possibly non-monophyletic) specified as a set of tips.

Usage

which.edge(phy, group)

Arguments

phy
an object of class "phylo".

group
a vector of mode numeric or character specifying the tips for which the edges are to be identified.

Details

The group of tips specified in ‘group’ may be non-monophyletic (paraphyletic or polyphyletic), in which case all edges from the tips to their most recent common ancestor are identified.

The identification is made with the indices of the rows of the matrix ‘edge’ of the tree.

Value

a numeric vector.

Author(s)

Emmanuel Paradis

See Also

bind.tree, drop.tip, root

[Package ape version 5.2 Index]