woodmouse {ape}	R Documentation

Cytochrome b Gene Sequences of Woodmice

Description

This is a set of 15 sequences of the mitochondrial gene cytochrome b of the woodmouse (Apodemus sylvaticus) which is a subset of the data analysed by Michaux et al. (2003). The full data set is available through GenBank (accession numbers AJ511877 to AJ511987).

Usage

data(woodmouse)

Format

An object of class "DNAbin".

Source

Michaux, J. R., Magnanou, E., Paradis, E., Nieberding, C. and Libois, R. (2003) Mitochondrial phylogeography of the Woodmouse (Apodemus sylvaticus) in the Western Palearctic region. Molecular Ecology, 12, 685–697.

See Also

read.dna, DNAbin, dist.dna

Examples

data(woodmouse)
str(woodmouse)

[Package ape version 5.2 Index]