write.tree {ape}	R Documentation

Write Tree File in Parenthetic Format

Description

This function writes in a file a tree in parenthetic format using the Newick (also known as New Hampshire) format.

Usage

write.tree(phy, file = "", append = FALSE,
           digits = 10, tree.names = FALSE)

Arguments

phy
an object of class "phylo" or "multiPhylo".

file
a file name specified by either a variable of mode character, or a double-quoted string; if file = "" (the default) then the tree is written on the standard output connection (i.e. the console).

append
a logical, if TRUE the tree is appended to the file without erasing the data possibly existing in the file, otherwise the file (if it exists) is overwritten (FALSE the default).

digits
a numeric giving the number of digits used for printing branch lengths.

tree.names
either a logical or a vector of mode character. If TRUE then any tree names will be written prior to the tree on each line. If character, specifies the name of "phylo" objects which can be written to the file.

Details

The node labels and the root edge length, if available, are written in the file.

If tree.names == TRUE then a variant of the Newick format is written for which the name of a tree precedes the Newick format tree (parentheses are eventually deleted beforehand). The tree names are taken from the names attribute if present (they are ignored if tree.names is a character vector).

The tip labels (and the node labels if present) are checked before being printed: the leading and trailing spaces, and the leading left and trailing right parentheses are deleted; the other spaces are replaced by underscores; the commas, colons, semicolons, and the other parentheses are replaced with dashes.

Value

a vector of mode character if file = "", none (invisible NULL) otherwise.

Author(s)

Emmanuel Paradis, Daniel Lawson dan.lawson@bristol.ac.uk, and Klaus Schliep kschliep@snv.jussieu.fr

References

Felsenstein, J. The Newick tree format. http://evolution.genetics.washington.edu/phylip/newicktree.html

Olsen, G. Interpretation of the "Newick's 8:45" tree format standard. http://evolution.genetics.washington.edu/phylip/newick_doc.html

See Also

read.tree, read.nexus, write.nexus

[Package ape version 5.2 Index]