SDM {ape}	R Documentation

Construction of Consensus Distance Matrix With SDM

Description

This function implements the SDM method of Criscuolo et al. (2006) for a set of n distance matrices.

Usage


SDM(...)

Arguments

...
2n elements (with n > 1), the first n elements are the distance matrices: these can be (symmetric) matrices, objects of class "dist", or a mix of both. The next n elements are the sequence length from which the matrices have been estimated (can be seen as a degree of confidence in matrices).

Details

Reconstructs a consensus distance matrix from a set of input distance matrices on overlapping sets of taxa. Potentially missing values in the supermatrix are represented by NA. An error is returned if the input distance matrices can not resolve to a consensus matrix.

Value

a 2-element list containing a distance matrix labelled by the union of the set of taxa of the input distance matrices, and a variance matrix associated to the returned distance matrix.

Author(s)

Andrei Popescu

References

Criscuolo, A., Berry, V., Douzery, E. J. P. , and Gascuel, O. (2006) SDM: A fast distance-based approach for (super)tree building in phylogenomics. Systematic Biology, 55, 740–755.

See Also

bionj, fastme, njs, mvrs, triangMtd

[Package ape version 5.5 Index]