alview {ape}	R Documentation

Print DNA or AA Sequence Alignement

Description

This function displays in the console or a file an alignment of DNA or AAsequences. The first sequence is printed on the first row and the bases of the other sequences are replaced by dots if they are identical with the first sequence.

Usage


alview(x, file = "", uppercase = TRUE, showpos = TRUE)

Arguments

x
a matrix or a list of DNA sequences (class "DNAbin") or a matrix of AA sequences (class "AAbin").

file
a character string giving the name of the file where to print the sequences; by default, they are printed in the console.

uppercase
a logical specifying whether to print the bases as uppercase letters.

showpos
either a logical value specifying whether to display the site positions, or a numeric vector giving these positions (see examples).

Details

The first line of the output shows the position of the last column of the printed alignment.

Author(s)

Emmanuel Paradis

See Also

DNAbin, image.DNAbin, alex, clustal, checkAlignment, all.equal.DNAbin

Examples


data(woodmouse)
alview(woodmouse[, 1:50])
alview(woodmouse[, 1:50], uppercase = FALSE)
## display only some sites:
j <- c(10, 49, 125, 567) # just random
x <- woodmouse[, j]
alview(x, showpos = FALSE) # no site position displayed
alview(x, showpos = j)
## Not run: 
alview(woodmouse, file = "woodmouse.txt")

## End(Not run)

[Package ape version 5.5 Index]