as.bitsplits {ape}	R Documentation

Split Frequencies and Conversion Among Split Classes

Description

bitsplits returns the bipartitions (aka splits) for a single tree or a list of trees. If at least one tree is rooted, an error is returned.

countBipartitions returns the frequencies of the bipartitions from a reference tree (phy) observed in a list of trees (X), all unrooted.

as.bitsplits and as.prop.part are generic functions for converting between the "bitsplits" and "prop.part" classes.

Usage


bitsplits(x)
countBipartitions(phy, X)
as.bitsplits(x)
## S3 method for class 'prop.part'
as.bitsplits(x)
## S3 method for class 'bitsplits'
print(x, ...)
## S3 method for class 'bitsplits'
sort(x, decreasing = FALSE, ...)
as.prop.part(x, ...)
## S3 method for class 'bitsplits'
as.prop.part(x, include.trivial = FALSE, ...)

Arguments

x
an object of the appropriate class.

phy
an object of class "phylo".

X
an object of class "multiPhylo".

decreasing
a logical value to sort the bipartitions in increasing (the default) or decreasing order of their frequency.

include.trivial
a logical value specifying whether to include the trivial split with all tips in the returned object.

...
further arguments passed to or from other methods.

Details

These functions count bipartitions as defined by internal branches, so they not work only with unrooted trees. The structure of the class "bitsplits" is described in a separate document on ape's web site.

Value

bitsplits, as.bitsplits, and sort return an object of class "bitsplits".

countBipartitions returns a vector of integers.

as.prop.part returns an object of class "prop.part".

Author(s)

Emmanuel Paradis

See Also

prop.part, is.compatible

Examples


tr <- rtree(20)
pp <- prop.part(tr)
as.bitsplits(pp)
## works only with unrooted trees (ape 5.5):
countBipartitions(rtree(10, rooted = FALSE), rmtree(100, 10, rooted = FALSE))

[Package ape version 5.5 Index]