corBlomberg {ape}	R Documentation

Blomberg et al.'s Correlation Structure

Description

The “ACDC” (accelerated/decelerated) model assumes that continuous traits evolve under a Brownian motion model which rates accelerates (if g < 1) or decelerates (if g > 1) through time. If g = 1, then the model reduces to a Brownian motion model.

Usage


corBlomberg(value, phy, form = ~1, fixed = FALSE)
## S3 method for class 'corBlomberg'
corMatrix(object, covariate = getCovariate(object),
                   corr = TRUE, ...)
## S3 method for class 'corBlomberg'
coef(object, unconstrained = TRUE, ...)

Arguments

value
the (initial) value of the parameter g.

phy
an object of class "phylo".

form
a one sided formula of the form ~ t, or ~ t | g, specifying the taxa covariate t and, optionally, a grouping factor g. A covariate for this correlation structure must be character valued, with entries matching the tip labels in the phylogenetic tree. When a grouping factor is present in form, the correlation structure is assumed to apply only to observations within the same grouping level; observations with different grouping levels are assumed to be uncorrelated. Defaults to ~ 1, which corresponds to using the order of the observations in the data as a covariate, and no groups.

fixed
a logical specifying whether gls should estimate gamma (the default) or keep it fixed.

object
an (initialized) object of class "corBlomberg".

covariate
an optional covariate vector (matrix), or list of covariate vectors (matrices), at which values the correlation matrix, or list of correlation matrices, are to be evaluated. Defaults to getCovariate(object).

corr
a logical value specifying whether to return the correlation matrix (the default) or the variance-covariance matrix.

unconstrained
a logical value. If TRUE (the default), the coefficients are returned in unconstrained form (the same used in the optimization algorithm). If FALSE the coefficients are returned in “natural”, possibly constrained, form.

...
further arguments passed to or from other methods.

Value

an object of class "corBlomberg", the coefficients from an object of this class, or the correlation matrix of an initialized object of this class. In most situations, only corBlomberg will be called by the user.

Author(s)

Emmanuel Paradis

References

Blomberg, S. P., Garland, Jr, T., and Ives, A. R. (2003) Testing for phylogenetic signal in comparative data: behavioral traits are more labile. Evolution, 57, 717–745.

[Package ape version 5.5 Index]