corPagel {ape}	R Documentation

Pagel's “lambda” Correlation Structure

Description

The correlation structure from the present model is derived from the Brownian motion model by multiplying the off-diagonal elements (i.e., the covariances) by lambda. The variances are thus the same than for a Brownian motion model.

Usage


corPagel(value, phy, form = ~1, fixed = FALSE)
## S3 method for class 'corPagel'
corMatrix(object, covariate = getCovariate(object),
                   corr = TRUE, ...)
## S3 method for class 'corPagel'
coef(object, unconstrained = TRUE, ...)

Arguments

value
the (initial) value of the parameter lambda.

phy
an object of class "phylo".

form
a one sided formula of the form ~ t, or ~ t | g, specifying the taxa covariate t and, optionally, a grouping factor g. A covariate for this correlation structure must be character valued, with entries matching the tip labels in the phylogenetic tree. When a grouping factor is present in form, the correlation structure is assumed to apply only to observations within the same grouping level; observations with different grouping levels are assumed to be uncorrelated. Defaults to ~ 1, which corresponds to using the order of the observations in the data as a covariate, and no groups.

fixed
a logical specifying whether gls should estimate lambda (the default) or keep it fixed.

object
an (initialized) object of class "corPagel".

covariate
an optional covariate vector (matrix), or list of covariate vectors (matrices), at which values the correlation matrix, or list of correlation matrices, are to be evaluated. Defaults to getCovariate(object).

corr
a logical value specifying whether to return the correlation matrix (the default) or the variance-covariance matrix.

unconstrained
a logical value. If TRUE (the default), the coefficients are returned in unconstrained form (the same used in the optimization algorithm). If FALSE the coefficients are returned in “natural”, possibly constrained, form.

...
further arguments passed to or from other methods.

Value

an object of class "corPagel", the coefficients from an object of this class, or the correlation matrix of an initialized object of this class. In most situations, only corPagel will be called by the user.

Author(s)

Emmanuel Paradis

References

Freckleton, R. P., Harvey, P. H. and M. Pagel, M. (2002) Phylogenetic analysis and comparative data: a test and review of evidence. American Naturalist, 160, 712–726.

Pagel, M. (1999) Inferring the historical patterns of biological evolution. Nature, 401,877–884.

[Package ape version 5.5 Index]