hivtree {ape}	R Documentation

Phylogenetic Tree of 193 HIV-1 Sequences

Description

This data set describes an estimated clock-like phylogeny of 193 HIV-1 group M sequences sampled in the Democratic Republic of Congo.

Usage

data(hivtree.newick)
data(hivtree.table)

Format

hivtree.newick is a string with the tree in Newick format. The data frame hivtree.table contains the corresponding internode distances.

Source

This is a data example from Strimmer and Pybus (2001).

References

Strimmer, K. and Pybus, O. G. (2001) Exploring the demographic history of DNA sequences using the generalized skyline plot. Molecular Biology and Evolution, 18, 2298–2305.

See Also

coalescent.intervals, collapsed.intervals

[Package ape version 5.8 Index]