njs {ape}	R Documentation

Tree Reconstruction from Incomplete Distances With NJ* or bio-NJ*

Description

Reconstructs a phylogenetic tree from a distance matrix with possibly missing values.

Usage


njs(X, fs = 15)
bionjs(X, fs = 15)

Arguments

X
a distance matrix.

fs
argument s of the agglomerative criterion: it is coerced as an integer and must at least equal to one.

Details

Missing values represented by either NA or any negative number.

Basically, the Q* criterion is applied to all the pairs of leaves, and the s highest scoring ones are chosen for further analysis by the agglomeration criteria that better handle missing distances (see references for details).

Value

an object of class "phylo".

Author(s)

Andrei Popescu

References

Criscuolo, A., Gascuel, O. (2008) Fast NJ-like algorithms to deal with incomplete distance matrices. BMC Bioinformatics, 9, 166.

See Also

nj, bionj, triangMtds

Examples


data(woodmouse)
d <- dist.dna(woodmouse)
dm <- d
dm[sample(length(dm), size = 3)] <- NA
dist.topo(njs(dm), nj(d)) # often 0
dm[sample(length(dm), size = 10)] <- NA
dist.topo(njs(dm), nj(d)) # sometimes 0

[Package ape version 5.5 Index]